fix(mfa): widen digits=10 modulus/entropy to uint64 (review finding, Critical)
Independent review caught a silent correctness bug: HOTPCode's and
GenerateOTPCode's decimal-modulus accumulator (mod *= 10 in a loop) was
uint32, but 10^10 (10000000000) exceeds uint32's max (4294967295) — so
digits=10, the documented maximum this package advertises supporting, wrapped
to an effective modulus of ~1.41e9 and silently produced codes confined to a
much smaller range than 10 digits actually spans, with no error returned.
GenerateOTPCode also only read 4 random bytes (32 bits), too little entropy
to reduce mod 10^10 without a pronounced bias even after fixing the modulus
width.

Fix: both HOTPCode's dynamic-truncation binCode/mod and GenerateOTPCode's
mod/random draw are now uint64 (GenerateOTPCode reads 8 random bytes).
Digits 1-9 were never affected (their moduli fit uint32); only the digits=10
boundary was wrong.

Added TestHOTP_MaxDigitsSpansFullRange and TestOTP_MaxDigitsSpansFullRange,
which assert a digits=10 code reaches into the range the uint32-wrapped
modulus could never produce; both verified to fail against the pre-fix code
(reverted locally to confirm) and pass against the fix.

kernel/mfa coverage unchanged at 97.8%; make lint-new/test-unit/coverage-check
all still pass (repo-wide 91.8%, floor 90%).

+++ kernel/mfa/hotp.go
@@ -68,14 +68,20 @@ func HOTPCode(key []byte, counter uint64, alg Algorithm, digits int) (string, er
 	mac.Write(counterBytes[:])
 	sum := mac.Sum(nil)
 
-	// RFC 4226 §5.3 dynamic truncation.
+	// RFC 4226 §5.3 dynamic truncation. binCode is bounded by the 31-bit mask
+	// on the top byte (max 2147483647), which needs up to 10 decimal digits —
+	// so the modulus below MUST be computed in a width wider than uint32:
+	// 10^10 (10000000000) already exceeds uint32's max (4294967295), and a
+	// uint32 accumulator would silently wrap to a much smaller modulus,
+	// truncating the code's range without ever returning an error. uint64
+	// comfortably holds 10^10.
 	offset := sum[len(sum)-1] & 0x0f
-	binCode := (uint32(sum[offset])&0x7f)<<24 |
-		(uint32(sum[offset+1])&0xff)<<16 |
-		(uint32(sum[offset+2])&0xff)<<8 |
-		(uint32(sum[offset+3]) & 0xff)
+	binCode := (uint64(sum[offset])&0x7f)<<24 |
+		(uint64(sum[offset+1])&0xff)<<16 |
+		(uint64(sum[offset+2])&0xff)<<8 |
+		(uint64(sum[offset+3]) & 0xff)
 
-	mod := uint32(1)
+	mod := uint64(1)
 	for i := 0; i < digits; i++ {
 		mod *= 10
 	}
